Add style in Propriétaire page .
