+++ src/pages/articleDetails/ArticleDetails.scss
@@ -1,0 +1,23 @@
+#details-container{
+    // background-color: red;
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    margin: 10px 40px;
+    
+
+    img {
+        width: auto;
+        height: 500px;
+        border-radius: 20px;
+    }
+
+    h1 {
+        font-size: 1.5rem;
+        font-weight: bold;
+    }
+
+    li{
+        list-style:disc;
+    }
+}
+
